Trying to get save method to work

+++ start_code/db/cinema.sql
@@ -5,14 +5,14 @@ DROP TABLE film;
 CREATE TABLE film(
   id SERIAL PRIMARY KEY,
   title VARCHAR(255),
-  price INT,
-  budget INT
+  price NUMERIC,
+  budget NUMERIC
 );
 
 CREATE TABLE customer(
   id SERIAL PRIMARY KEY,
   name VARCHAR(255),
-  funds INT
+  funds NUMERIC
 );
 
 -- CREATE TABLE ticket(
